+++ src/cltkv1/utils/pipelines.py
@@ -9,8 +9,8 @@ these dataclasses is to represent:
 from dataclasses import dataclass
 from typing import Callable, List
 
+from cltkv1 import DefaultTokenizationOperation, LatinTokenizationOperation
 from cltkv1.languages.glottolog import LANGUAGES
-from cltkv1.tokenizers import DefaultTokenizationOperation, LatinTokenizationOperation
 from cltkv1.utils.data_types import Doc, Word
